First attempt to fix BLE bug (#71)
* First attempt to fix BLE bug

This commit changes the removes @abandonware/bluetooth-hci-socket
package and updates both bleno and noble to their latest version as the
hci socket library is not really necessary with the newer versions.

Furthermore, this commeit removes the concurrent central and peripheral
hack as well as switches to the default noble initialization.

* Second attempt to fix the BLE bug

This commit refactors the whole BLE HRM service to use async/await
that better streamlines the connection flow. It removes the listeners
on the characteristics subscription to avoid potential leaks.

Also adds NOBLE_MULTI_ROLE=1 env var to the npm start command.

* Change the BLE library

Move the BLE library to stoprocent/bleno and noble

---------

Co-authored-by: Abász <>